+++ src/sql/src/pure.rs
@@ -71,7 +71,6 @@ pub fn purify(
     };
 
     let now = catalog.now();
-    let experimental_mode = catalog.config().experimental_mode;
 
     async move {
         if let Statement::CreateSource(CreateSourceStatement {
@@ -176,13 +175,6 @@ pub fn purify(
                     publication,
                     slot,
                 } => {
-                    if !experimental_mode {
-                        bail!(
-                            "Postgres sources require experimental mode; \
-                            see https://materialize.com/docs/cli/#experimental-mode"
-                        );
-                    }
-
                     slot.get_or_insert_with(|| {
                         format!(
                             "materialize_{}",